requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
creyPY==1.2.5
fastapi==0.115.0
fastapi-pagination==0.12.31
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2

Mako==1.3.5 # Alembic
MarkupSafe==3.0.1 # Alembic
alembic==1.13.3 # Alembic

SQLAlchemy-Utils==0.41.2 # SQLAlchemy

click==8.1.7 # Uvicorn
uvicorn==0.31.1 # Uvicorn

iniconfig==2.0.0 # pytest
packaging==24.1 # pytest
pluggy==1.5.0 # pytest
pytest==8.3.3 # pytest
